pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cluspro-automation-py"
version = "0.4.0"
description = "Python automation tool for ClusPro protein docking web server"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "E Wijaya"}
]
keywords = ["cluspro", "protein-docking", "automation", "selenium", "bioinformatics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
validate = [
    "biopython>=1.80",
    "scipy>=1.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "types-beautifulsoup4>=4.12.0",
    "types-requests>=2.31.0",
]
all = [
    "cluspro-automation-py[validate,dev]",
]

[project.scripts]
cluspro = "cluspro.cli:main"

[project.urls]
Homepage = "https://github.com/ewijaya/cluspro-automation-py"
Documentation = "https://github.com/ewijaya/cluspro-automation-py#readme"
Issues = "https://github.com/ewijaya/cluspro-automation-py/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N812"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
